+++ src/core/cpu.hpp
@@ -12,7 +12,7 @@ class CPU {
   public:
     explicit CPU(Config &cfg);
 
-    void load_program(const std::vector<uint8_t> &code, uint32_t address = 0);
+    void load_program(const std::vector<uint8_t> &code, uint64_t address = 0);
     void step();
     void step_uop();
     void run(int max_cycles = 10000);
@@ -27,7 +27,7 @@ class CPU {
 
     bool is_halted() const { return executor_.is_halted(); }
     const std::vector<uint8_t> &get_code() const { return code_; }
-    uint32_t get_load_address() const { return load_address_; }
+    uint64_t get_load_address() const { return load_address_; }
 
     std::vector<DeclarativePeripheral> &get_peripherals() {
         return dec_peripherals_;
@@ -42,6 +42,6 @@ class CPU {
     Executor executor_;
 
     std::vector<uint8_t> code_;
-    uint32_t load_address_;
+    uint64_t load_address_;
     std::vector<DeclarativePeripheral> dec_peripherals_;
 };
